Updates for simulation with delays.
Added a backward linear lookup operation: used when a message arrives
after a merge and must be send to the father.
Send no merge operation: inform a son that the merge won't continue.
Retry: if a lookup can not continue due to a non-stable PhtNode, send
a retry.
A map of all PhtNode (label, NodeId) has been added: useful to check
whether the dht routing is correct or not.